명언 박스 오른쪽에 원칙 뜨게
